use only one eval model, temporally disable softprompt eval
