--- nemo/collections/nlp/models/language_modeling/megatron_gpt_eval_model.py
@@ -1,30 +1,0 @@
-# Copyright (c) 2021, NVIDIA CORPORATION.  All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from typing import Any, Optional
-
-from omegaconf import OmegaConf
-
-from nemo.collections.nlp.models.language_modeling.megatron_gpt_model import MegatronGPTModel
-from nemo.collections.nlp.modules.common.text_generation_utils import generate
-
-__all__ = ["MegatronGPTEvalModel"]
-
-
-class MegatronGPTEvalModel(MegatronGPTModel):
-    def predict_step(self, batch: Any, batch_idx: int, dataloader_idx: Optional[int] = None) -> Any:
-        config = self.cfg.inference
-        config = OmegaConf.to_container(config)
-        config['sentences'] = batch
-        return generate(self, **config)
