+++ nemo/collections/nlp/models/language_modeling/megatron_gpt_model.py
@@ -17,7 +17,7 @@ import re
 from typing import Any, Dict, List, Optional, Union
 
 import torch
-import torch.nn.functional as F
+from omegaconf import OmegaConf
 from omegaconf.dictconfig import DictConfig
 from omegaconf.omegaconf import open_dict
 from pytorch_lightning.plugins.precision.native_amp import NativeMixedPrecisionPlugin
@@ -34,10 +34,8 @@ from nemo.collections.nlp.models.nlp_model import NLPModel
 from nemo.collections.nlp.modules.common.megatron.clip_grads import clip_grad_norm_fp32
 from nemo.collections.nlp.modules.common.megatron.megatron_init import initialize_model_parallel_for_nemo
 from nemo.collections.nlp.modules.common.megatron.module import Float16Module
-from nemo.collections.nlp.modules.common.megatron.utils import (
-    average_losses_across_data_parallel_group,
-    get_ltor_masks_and_position_ids,
-)
+from nemo.collections.nlp.modules.common.megatron.utils import average_losses_across_data_parallel_group
+from nemo.collections.nlp.modules.common.text_generation_utils import generate
 from nemo.collections.nlp.modules.common.tokenizer_utils import get_nmt_tokenizer
 from nemo.collections.nlp.parts.nlp_overrides import GradScaler
 from nemo.collections.nlp.parts.utils_funcs import get_last_rank
@@ -45,7 +43,7 @@ from nemo.core.optim import MainParamsOptimizerWrapper, prepare_lr_scheduler
 from nemo.utils import AppState, logging
 
 try:
-    from apex.transformer import parallel_state, tensor_parallel
+    from apex.transformer import parallel_state
     from apex.contrib.layer_norm.layer_norm import FastLayerNorm
     from apex.normalization.fused_layer_norm import FusedLayerNorm  # NOQA
     from apex.transformer.pipeline_parallel.schedules.common import (
@@ -56,7 +54,7 @@ try:
         forward_backward_pipelining_without_interleaving,
     )
     from apex.transformer.pipeline_parallel.schedules.fwd_bwd_no_pipelining import forward_backward_no_pipelining
-    from apex.transformer.pipeline_parallel.utils import get_num_microbatches, _reconfigure_microbatch_calculator
+    from apex.transformer.pipeline_parallel.utils import get_num_microbatches
 
     HAVE_APEX = True
 except (ImportError, ModuleNotFoundError):
@@ -912,270 +910,10 @@ class MegatronGPTModel(NLPModel):
         return params
 
     def predict_step(self, batch: Any, batch_idx: int, dataloader_idx: Optional[int] = None) -> Any:
-        request, positions, tokens_to_generate, compute_logprobs = MegatronGPTModel._bucketize_gpt_inference(
-            batch, self.use_soft_prompts
-        )
-
-        if compute_logprobs:
-            response = self.compute_logprobs(request, positions)
-        else:
-            response = self.complete(request, positions, tokens_to_generate)
-
-        return response
-
-    def complete(self, request: Dict, positions: List, tokens_to_generate: int):
-        """
-            Autoregressively invokes language model in the inference mode
-        Args:
-            request:
-                * tokens: List of "buckets" with unpadded tokens of the same length
-                * prompt_tags: List of "buckets" where each bucket contains the prompt_tag strings
-                               specifying the prompt tag to use (optional)
-            positions: List with initial prompts positions
-            tokens_to_generate: int value denoting amount of tokens model should generate
-
-        Returns:
-            response: A python list of tuples
-                (text, tokens, log_probs, offsets)
-                * text: string, inputted prompt + generated text by model
-                * tokens: list of tokens correspond to text
-                * log_probs: list of tokens log probabilities
-                * offsets: list of tokens start positions in text
-
-        """
-        app_state = AppState()
-
-        results = []
-        request_tokens = request["tokens"]
-
-        for idx, tokens in enumerate(request_tokens):
-            micro_batch_size = tokens.shape[0]
-            _reconfigure_microbatch_calculator(
-                rank=app_state.global_rank,
-                rampup_batch_size=None,
-                global_batch_size=micro_batch_size,
-                micro_batch_size=micro_batch_size,
-                data_parallel_size=1,
-            )
-
-            # For prompt tuned GPT models
-            if self.use_soft_prompts:
-                if self.cfg.get('pipeline_model_parallel_size', 1) > 1:
-                    raise ValueError('complete method is not yet supported for pipeline with soft prompts')
-                prompt_tags = request["prompt_tags"][idx]
-                prompt_tags_to_ids = dict(self.prompt_table)
-                prompt_ids = torch.tensor([prompt_tags_to_ids[tag] for tag in prompt_tags])
-            else:
-                prompt_ids = None
-
-            logsoftmaxlayer = torch.nn.LogSoftmax(dim=-1)
-
-            for i in range(tokens_to_generate + 1):
-                if self.use_soft_prompts:
-                    batch_size = len(tokens)
-                    full_length = len(tokens[0]) + self.num_prompt_tokens
-
-                    # Get postion ids for text after soft prompt
-                    position_ids = torch.arange(
-                        start=self.num_prompt_tokens, end=full_length, dtype=torch.long, device=self.device
-                    )
-                    position_ids = position_ids.unsqueeze(0).expand_as(tokens).clone()
-
-                    # Make attention mask starting with first token in soft prompt
-                    attention_mask = torch.tril(
-                        torch.ones((batch_size, full_length, full_length), device=self.device)
-                    ).view(batch_size, 1, full_length, full_length)
-                    attention_mask = attention_mask < 0.5
-
-                else:
-                    attention_mask, _, position_ids = get_ltor_masks_and_position_ids(
-                        data=tokens,
-                        eod_token=self.tokenizer.eos_id,
-                        reset_position_ids=self.cfg.get('reset_position_ids', False),
-                        reset_attention_mask=self.cfg.get('reset_attention_mask', False),
-                        eod_mask_loss=self.cfg.get('eod_mask_loss', False),
-                    )
-                attention_mask_repeat = torch.concat([attention_mask for _ in range(micro_batch_size)])
-                if self.use_soft_prompts:
-                    batch = [tokens, attention_mask_repeat, position_ids, prompt_ids]
-                else:
-                    batch = [tokens, attention_mask_repeat, position_ids]
-                tensor_shape = [tokens.shape[1], micro_batch_size, self.cfg.hidden_size]
-                if self.cfg.get('pipeline_model_parallel_size', 1) > 1:
-                    output_tensor = forward_backward_pipelining_without_interleaving(
-                        forward_step_func=self.get_forward_output_only_func(),
-                        batch=batch,
-                        model=self.model,
-                        forward_only=True,
-                        tensor_shape=tensor_shape,
-                        dtype=self.autocast_dtype,
-                    )
-                else:
-                    output_tensor = forward_backward_no_pipelining(
-                        forward_step_func=self.get_forward_output_only_func(),
-                        batch=batch,
-                        model=self.model,
-                        forward_only=True,
-                        tensor_shape=tensor_shape,
-                        dtype=self.autocast_dtype,
-                    )
-
-                # get output tensor
-                if parallel_state.is_pipeline_last_stage():
-                    output_tensor = output_tensor[0]['logits']
-                    output_tensor = tensor_parallel.gather_from_tensor_model_parallel_region(output_tensor)
-
-                    log_probs, token_ids = torch.max(logsoftmaxlayer(output_tensor), dim=-1)
-                    tokens = torch.cat([tokens, torch.unsqueeze(token_ids[:, -1], 1)], dim=1)
-                else:
-                    log_probs = torch.zeros((tokens.shape[0], tokens.shape[1]), dtype=torch.float).cuda()
-                    tokens = torch.zeros((tokens.shape[0], tokens.shape[1] + 1), dtype=tokens.dtype).cuda()
-
-                torch.distributed.broadcast(tokens, get_last_rank())
-                torch.distributed.broadcast(log_probs, get_last_rank())
-
-            # add to results as (text, tokens, log_probs, offsets)
-            for token, prob in zip(tokens, log_probs.tolist()):
-                results.append(
-                    (self.tokenizer.ids_to_text(token[:-1]), self.tokenizer.ids_to_tokens(token[:-1]), prob, [0],)
-                )
-
-        # offsets calculation
-        for item in results:
-            for index, token in enumerate(item[1]):
-                if index != len(item[1]) - 1:
-                    item[3].append(len(token) + item[3][-1])
-
-        # return prompts in the order that they were input
-        response = [0 for i in range(len(positions))]
-        for item, index in zip(results, positions):
-            response[index] = item
-
-        return response
-
-    def compute_logprobs(self, request: Dict, positions: List):
-        """
-            Only logprobs computation without generation tokens
-        Args:
-            request:
-                * tokens: List of "buckets" with unpadded tokens of the same length
-                * prompt_tags: List of "buckets" where each bucket contains the prompt_tag strings
-                                    specifying the prompt tag to use (optional)
-            positions: List with initial prompts positions
-        Returns:
-            response: A python list of tuples
-            (text, tokens, log_probs, offsets)
-            * text: string, inputted prompt + generated text by model
-            * tokens: list of tokens correspond to text
-            * log_probs: list of log_softmax's from output_tensor in respect to text tokens
-            * offsets: list of tokens start positions in text
-        """
-        app_state = AppState()
-
-        results = []
-        request_tokens = request["tokens"]
-        for idx, tokens in enumerate(request_tokens):
-            tokens_cut = tokens[:, :-1]
-            micro_batch_size = tokens_cut.shape[0]
-            _reconfigure_microbatch_calculator(
-                rank=app_state.global_rank,
-                rampup_batch_size=None,
-                global_batch_size=micro_batch_size,
-                micro_batch_size=micro_batch_size,
-                data_parallel_size=1,
-            )
-            # For prompt tuned GPT models
-            if self.use_soft_prompts:
-                if self.cfg.get('pipeline_model_parallel_size', 1) > 1:
-                    raise ValueError('compute_logprobs method is not yet supported for pipeline with soft prompts')
-                prompt_tags = request["prompt_tags"][idx]
-                prompt_tags_to_ids = dict(self.prompt_table)
-                prompt_ids = torch.tensor([prompt_tags_to_ids[tag] for tag in prompt_tags])
-            else:
-                prompt_ids = None
-
-            if self.use_soft_prompts:
-                batch_size = len(tokens_cut)
-                full_length = len(tokens_cut[0]) + self.num_prompt_tokens
-                # Get postion ids for text after soft prompt
-                position_ids = torch.arange(
-                    start=self.num_prompt_tokens, end=full_length, dtype=torch.long, device=self.device
-                )
-                position_ids = position_ids.unsqueeze(0).expand_as(tokens_cut).clone()
-                # Make attention mask starting with first token in soft prompt
-                attention_mask = torch.tril(
-                    torch.ones((batch_size, full_length, full_length), device=self.device)
-                ).view(batch_size, 1, full_length, full_length)
-                attention_mask = attention_mask < 0.5
-
-            else:
-                attention_mask, _, position_ids = get_ltor_masks_and_position_ids(
-                    data=tokens_cut,
-                    eod_token=self.tokenizer.eos_id,
-                    reset_position_ids=self.cfg.get('reset_position_ids', False),
-                    reset_attention_mask=self.cfg.get('reset_attention_mask', False),
-                    eod_mask_loss=self.cfg.get('eod_mask_loss', False),
-                )
-
-            # we repeat attention mask to work with apex fwd/bwd function
-            attention_mask_repeat = torch.concat([attention_mask for _ in range(micro_batch_size)])
-            if self.use_soft_prompts:
-                batch = [tokens_cut, attention_mask_repeat, position_ids, prompt_ids]
-            else:
-                batch = [tokens_cut, attention_mask_repeat, position_ids]
-            tensor_shape = [tokens_cut.shape[1], micro_batch_size, self.cfg.hidden_size]
-            if self.cfg.get('pipeline_model_parallel_size', 1) > 1:
-                output_tensor = forward_backward_pipelining_without_interleaving(
-                    forward_step_func=self.get_forward_output_only_func(),
-                    batch=batch,
-                    model=self.model,
-                    forward_only=True,
-                    tensor_shape=tensor_shape,
-                    dtype=self.autocast_dtype,
-                )
-            else:
-                output_tensor = forward_backward_no_pipelining(
-                    forward_step_func=self.get_forward_output_only_func(),
-                    batch=batch,
-                    model=self.model,
-                    forward_only=True,
-                    tensor_shape=tensor_shape,
-                    dtype=self.autocast_dtype,
-                )
-
-            # get output tensor
-            if parallel_state.is_pipeline_last_stage():
-                output_tensor = output_tensor[0]['logits']
-                output_tensor = tensor_parallel.gather_from_tensor_model_parallel_region(output_tensor)
-
-            else:
-                output_tensor = torch.zeros(
-                    (tokens_cut.shape[0], tokens_cut.shape[1], self.padded_vocab_size), dtype=torch.float
-                ).cuda()
-
-            torch.distributed.broadcast(output_tensor, get_last_rank())
-
-            log_probs = []
-            for output in output_tensor:
-                probs = F.log_softmax(output, dim=1)
-                probs = probs[-len(tokens_cut[0]) :]
-                log_probs.append(probs)
-
-            for token, prob in zip(tokens, log_probs):
-                results.append((self.tokenizer.ids_to_text(token), self.tokenizer.ids_to_tokens(token), prob, [0]))
-
-        # offsets calculation
-        for item in results:
-            for index, token in enumerate(item[1]):
-                if index != len(item[1]) - 1:
-                    item[3].append(len(token) + item[3][-1])
-
-        # return prompts in order they were inputted
-        response = [0 for i in range(len(positions))]
-        for item, index in zip(results, positions):
-            response[index] = item
-
-        return response
+        config = self.cfg.inference
+        config = OmegaConf.to_container(config)
+        config['sentences'] = batch
+        return generate(self, **config)
 
     def init_new_prompts(self):
         for idx, tag in enumerate(self.cfg.new_prompt_tags):
